refactor: Moved the creation of flux pulse sweepers to the resonator flux routine
