+++ src/qibocal/protocols/characterization/flux_dependence/resonator_crosstalk.py
@@ -14,6 +14,7 @@ from qibocal.protocols.characterization.flux_dependence.resonator_flux_dependenc
     ResFluxType,
     ResonatorFluxData,
     ResonatorFluxParameters,
+    create_flux_pulse_sweepers,
 )
 
 from . import utils
@@ -93,35 +94,11 @@ def _acquisition(
 
     else:
         flux_qubits = params.flux_qubits
-
+    drive_readout_duration = sequence.duration
     if params.flux_pulses:
-        qf_pulses = {}
-        for i, qubit in enumerate(qubits):
-            if isinstance(params.flux_amplitude_start, list):
-                flux_amplitude_start = params.flux_amplitude_start[i]
-                flux_amplitude_end = params.flux_amplitude_end[i]
-                flux_amplitude_step = params.flux_amplitude_step[i]
-            else:
-                flux_amplitude_start = params.flux_amplitude_start
-                flux_amplitude_end = params.flux_amplitude_end
-                flux_amplitude_step = params.flux_amplitude_step
-            delta_bias_flux_range = np.arange(
-                flux_amplitude_start,
-                flux_amplitude_end,
-                flux_amplitude_step,
-            )
-            pulse = platform.create_qubit_flux_pulse(
-                qubit, start=0, duration=sequence.duration
-            )
-            qf_pulses[qubit] = pulse
-        sweepers = [
-            Sweeper(
-                Parameter.amplitude,
-                delta_bias_flux_range,
-                pulses=[qf_pulses[flux_qubit] for flux_qubit in flux_qubits],
-                type=SweeperType.ABSOLUTE,
-            )
-        ]
+        sweepers = create_flux_pulse_sweepers(
+            params, platform, qubits, drive_readout_duration
+        )
     else:
         delta_bias_flux_range = np.arange(
             -params.bias_width / 2, params.bias_width / 2, params.bias_step
